Migrate more off of prototype code
