+++ src/sode/python/sode/soundcloud/auth/api.py
@@ -1,6 +1,7 @@
 import json
 import logging
 from dataclasses import asdict, dataclass
+from datetime import datetime, timedelta
 from typing import Any, Mapping, TextIO, Tuple
 
 from oauthlib.oauth2 import BackendApplicationClient
@@ -26,9 +27,36 @@ class TokenResponse:
     scope: list[str]  # ['']
     token_type: str  # Bearer
 
+    @staticmethod
+    def parse(data: Mapping[str, Any]) -> "TokenResponse":
+        """Parse data from JSON data like you get when fetching tokens with requests_oauthlib"""
+
+        return TokenResponse(**data)
+
     @property
     def access_token_v(self) -> Either[str, AccessToken]:
         return AccessToken.expected(self.access_token, self.token_type)
+
+    @property
+    def expiration(self) -> datetime:
+        """the datetime at which this access token expires"""
+
+        return datetime.fromtimestamp(self.expires_at)
+
+    @property
+    def expired(self, now: datetime = datetime.now()) -> bool:
+        """true if the access token has expired; false if not"""
+
+        return self.expiration > now
+
+    @property
+    def remaining_time(self, now: datetime = datetime.now()) -> timedelta:
+        """how much time is left before the access token expires, or 0 if it has already expired"""
+
+        if self.expiration <= now:
+            return timedelta()
+        else:
+            return self.expiration - now
 
     def write_json(
         self,
